More player page refactoring because why not
